# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: BodyLogic - AI-Powered Health & Fitness Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    get:
      operationId: getMeals
      tags: [nutrition]
      summary: Get logged meals
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: Filter by date (YYYY-MM-DD)
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: logMeal
      tags: [nutrition]
      summary: Log a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogMealBody"
      responses:
        "201":
          description: Meal logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    delete:
      operationId: deleteMeal
      tags: [nutrition]
      summary: Delete a meal log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Meal deleted
...
  /workouts:
    get:
      operationId: getWorkouts
      tags: [fitness]
      summary: Get logged workouts
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: logWorkout
      tags: [fitness]
      summary: Log a workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogWorkoutBody"
      responses:
        "201":
          description: Workout logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    delete:
      operationId: deleteWorkout
      tags: [fitness]
      summary: Delete a workout log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Workout deleted
...
  /sleep:
    get:
      operationId: getSleepLogs
      tags: [sleep]
      summary: Get sleep logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Sleep logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepLog"
    post:
      operationId: logSleep
      tags: [sleep]
      summary: Log sleep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogSleepBody"
      responses:
        "201":
          description: Sleep logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepLog"
...
components:
  schemas:
...
    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        vitamins:
          type: string
          nullable: true
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        loggedAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        mealIqScore:
          type: number
          nullable: true
      required: [id, name, calories, protein, carbs, fat, fiber, sugar, mealType, loggedAt]
    LogMealBody:
      type: object
      properties:
        name:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        vitamins:
          type: string
          nullable: true
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        notes:
          type: string
          nullable: true
      required: [name, calories, protein, carbs, fat, fiber, sugar, mealType]
...
    Workout:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [strength, cardio, flexibility, hiit, yoga, running, cycling, swimming, other]
        durationMinutes:
          type: integer
        caloriesBurned:
          type: integer
        intensity:
          type: string
          enum: [low, moderate, high]
        distanceKm:
          type: number
          nullable: true
        loggedAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required: [id, name, type, durationMinutes, caloriesBurned, intensity, loggedAt]
    LogWorkoutBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [strength, cardio, flexibility, hiit, yoga, running, cycling, swimming, other]
        durationMinutes:
          type: integer
        caloriesBurned:
          type: integer
        intensity:
          type: string
          enum: [low, moderate, high]
        distanceKm:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, type, durationMinutes, caloriesBurned, intensity]
...
    SleepLog:
      type: object
      properties:
        id:
          type: integer
        bedtime:
          type: string
          format: date-time
        wakeTime:
          type: string
          format: date-time
        durationHours:
          type: number
        quality:
          type: string
          enum: [poor, fair, good, excellent]
        deepSleepHours:
          type: number
        lightSleepHours:
          type: number
        remSleepHours:
          type: number
        awakeMins:
          type: integer
        sleepScore:
          type: integer
        date:
          type: string
          format: date
      required: [id, bedtime, wakeTime, durationHours, quality, deepSleepHours, lightSleepHours, remSleepHours, awakeMins, sleepScore, date]
    LogSleepBody:
      type: object
      properties:
        bedtime:
          type: string
          format: date-time
        wakeTime:
          type: string
          format: date-time
        quality:
          type: string
          enum: [poor, fair, good, excellent]
        deepSleepHours:
          type: number
        lightSleepHours:
          type: number
        remSleepHours:
          type: number
        awakeMins:
          type: integer
      required: [bedtime, wakeTime, quality]
...