# lib/api-zod/src/generated/types/sleepLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BodyLogic - AI-Powered Health & Fitness Tracker API
 * OpenAPI spec version: 0.2.0
 */
import type { SleepLogQuality } from "./sleepLogQuality";

export interface SleepLog {
  id: number;
  bedtime: Date;
  wakeTime: Date;
  durationHours: number;
  quality: SleepLogQuality;
  deepSleepHours: number;
  lightSleepHours: number;
  remSleepHours: number;
  awakeMins: number;
  sleepScore: number;
  date: Date;
}